# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "KCChangeApply": {
        "properties": {
          "confirm": {
            "type": "string"
          },
          "previewHash": {
            "type": "string"
          }
        },
        "required": [
          "previewHash",
          "confirm"
        ],
        "type": "object"
      },
      "KCChangeDecision": {
        "properties": {
          "decision": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "decision",
          "reason"
        ],
        "type": "object"
      },
      "KCChangeInput": {
        "properties": {
          "effectiveReleaseId": {
            "format": "uuid",
            "type": "string"
          },
          "fromKCIds": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "type": "array"
          },
          "fromReleaseId": {
            "format": "uuid",
            "type": "string"
          },
          "proposalType": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/KCChangeTarget"
            },
            "type": "array"
          }
        },
        "required": [
          "proposalType",
          "rationale",
          "fromReleaseId",
          "effectiveReleaseId",
          "fromKCIds",
          "targets"
        ],
        "type": "object"
      },
      "KCChangeProposal": {
        "properties": {
          "appliedAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "appliedBy": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "createdBy": {
            "format": "uuid",
            "type": "string"
          },
          "effectiveReleaseId": {
            "format": "uuid",
            "type": "string"
          },
          "familyId": {
            "format": "uuid",
            "type": "string"
          },
          "fromReleaseId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "proposalType": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "reviewedAt": {
            "format": "date-time",
            "type": [
              "null",
              "string"
            ]
          },
          "reviewedBy": {
            "format": "uuid",
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "type": "object"
      },
      "KCChangeProposalItem": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "familyId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kcId": {
            "format": "uuid",
            "type": "string"
          },
          "proposalId": {
            "format": "uuid",
            "type": "string"
          },
          "proposedRevisionId": {
            "format": "uuid",
            "type": "string"
          },
          "side": {
            "type": "string"
          },
          "weight": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          }
        },
        "type": "object"
      },
      "KCChangeTarget": {
        "properties": {
          "kcId": {
            "format": "uuid",
            "type": "string"
          },
          "weight": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          }
        },
        "required": [
          "kcId"
        ],
        "type": "object"
      },
      "KCProposalEvent": {
        "properties": {
          "action": {
            "type": "string"
          },
          "actorId": {
            "format": "uuid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "familyId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "proposalId": {
            "format": "uuid",
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "snapshot": {
            "type": "string"
          },
          "version": {
            "format": "int64",
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ]
          }
        },
        "type": "object"
      },
      "KnowledgeMigration": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "effectiveReleaseId": {
            "format": "uuid",
            "type": "string"
          },
          "evidencePolicy": {
            "type": "string"
          },
          "familyId": {
            "format": "uuid",
            "type": "string"
          },
          "fromKCId": {
            "format": "uuid",
            "type": "string"
          },
          "fromRevisionId": {
            "format": "uuid",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "migrationType": {
            "type": "string"
          },
          "proposalId": {
            "format": "uuid",
            "type": "string"
          },
          "toKCId": {
            "format": "uuid",
            "type": "string"
          },
          "toRevisionId": {
            "format": "uuid",
            "type": "string"
          },
          "weight": {
            "format": "double",
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d+)?$",
            "type": [
              "null",
              "number",
              "string"
            ]
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Learning.Api | v1",
    "version": "1.0.0"
  },
  "openapi": "3.1.1",
  "paths": {
...
    "/api/v1/content/kc-changes": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "identities": {
                      "items": {
                        "properties": {
                          "id": {
                            "format": "uuid",
                            "type": "string"
                          },
                          "status": {
                            "type": [
                              "null",
                              "string"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "status"
                        ],
                        "type": "object"
                      },
                      "type": [
                        "null",
                        "array"
                      ]
                    },
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/KCChangeProposalItem"
                      },
                      "type": [
                        "null",
                        "array"
                      ]
                    },
                    "migrations": {
                      "items": {
                        "$ref": "#/components/schemas/KnowledgeMigration"
                      },
                      "type": [
                        "null",
                        "array"
                      ]
                    },
                    "notice": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "proposals": {
                      "items": {
                        "$ref": "#/components/schemas/KCChangeProposal"
                      },
                      "type": [
                        "null",
                        "array"
                      ]
                    }
                  },
                  "required": [
                    "proposals",
                    "items",
                    "migrations",
                    "identities",
                    "notice"
                  ],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KCChangeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KCChangeProposal"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
    "/api/v1/content/kc-changes/{id}": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "events": {
                      "items": {
                        "$ref": "#/components/schemas/KCProposalEvent"
                      },
                      "type": [
                        "null",
                        "array"
                      ]
                    },
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/KCChangeProposalItem"
                      },
                      "type": [
                        "null",
                        "array"
                      ]
                    },
                    "migrations": {
                      "items": {
                        "$ref": "#/components/schemas/KnowledgeMigration"
                      },
                      "type": [
                        "null",
                        "array"
                      ]
                    },
                    "notice": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "proposal": {
                      "$ref": "#/components/schemas/KCChangeProposal"
                    }
                  },
                  "required": [
                    "proposal",
                    "items",
                    "events",
                    "migrations",
                    "notice"
                  ],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KCChangeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KCChangeProposal"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
    "/api/v1/content/kc-changes/{id}/preview": {
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "items": {
                      "items": {
                        "$ref": "#/components/schemas/KCChangeProposalItem"
                      },
                      "type": [
                        "null",
                        "array"
                      ]
                    },
                    "notice": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "previewHash": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "proposal": {
                      "$ref": "#/components/schemas/KCChangeProposal"
                    },
                    "requiredConfirm": {
                      "type": [
                        "null",
                        "string"
                      ]
                    }
                  },
                  "required": [
                    "proposal",
                    "items",
                    "previewHash",
                    "requiredConfirm",
                    "notice"
                  ],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
    "/api/v1/content/kc-changes/{id}:apply": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KCChangeApply"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "notice": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "proposal": {
                      "$ref": "#/components/schemas/KCChangeProposal"
                    }
                  },
                  "required": [
                    "proposal",
                    "notice"
                  ],
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
    "/api/v1/content/kc-changes/{id}:decide": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KCChangeDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KCChangeProposal"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
    "/api/v1/content/kc-changes/{id}:submit": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KCChangeProposal"
                }
              }
            }
          }
        },
        "tags": [
          "Learning.Api"
        ]
      }
    },
...
  },
...
}